Update main.cpp
`w.ws_row` has problems with different fonts, so sadly i have to add a number there which means that there's a task name limit (40 characters). I could truncate the name but there's no point in doing that if the user will never get to see the real name again.

If anyone knows a fix to this i'd appreciate it.